% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_agregations.R
\name{calculer_agregat_sur_croisement}
\alias{calculer_agregat_sur_croisement}
\title{Calculer l'agrégation sur croisement avec plusieurs fonctions}
\usage{
calculer_agregat_sur_croisement(
  table,
  var_croisement,
  var_croisement_relative = c(),
  var_quanti,
  liste_fonction_agregation,
  condition = NULL
)
}
\arguments{
\item{table}{Data.table contenant les données.}

\item{var_croisement}{Variables de croisement.}

\item{var_croisement_relative}{Variables de croisement relatives (optionnel).}

\item{var_quanti}{Variable quantitative à agréger.}

\item{liste_fonction_agregation}{Liste de fonctions d'agrégation (ex: list(sum, mean)).}

\item{condition}{Condition à appliquer (optionnelle).}
}
\value{
Un tibble avec les totaux et les parts pour chaque fonction d'agrégation.
}
\description{
Calculer l'agrégation sur croisement avec plusieurs fonctions
}
