% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_simple_calculer_agregations.R
\name{calculer_evolution_from_table_agrege}
\alias{calculer_evolution_from_table_agrege}
\title{Calculer l'évolution à partir d'une table agrégée}
\usage{
calculer_evolution_from_table_agrege(
  table_agrege,
  var_evolution,
  var_croisement,
  var_croisement_relative,
  var_quanti_selec
)
}
\arguments{
\item{table_agrege}{Un \code{data.frame} contenant les données agrégées.}

\item{var_evolution}{Un caractère spécifiant la variable d'évolution.}

\item{var_croisement}{Un vecteur de caractères spécifiant les variables de croisement.}

\item{var_croisement_relative}{Un vecteur de caractères spécifiant les variables de croisement relatives.}

\item{var_quanti}{Un caractère spécifiant la variable quantitative pour laquelle l'évolution est calculée.}
}
\value{
Un \code{data.frame} contenant les colonnes de croisement, les colonnes de croisement relative,
et la colonne d'évolution calculée.
}
\description{
Cette fonction calcule l'évolution d'une variable quantitative sur la base d'une table agrégée.
Elle transforme les données en un format large et calcule le pourcentage d'évolution entre les années.
}
