% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controler_demande.R
\name{controler_demandes}
\alias{controler_demandes}
\title{Contrôler la validité des demandes d'agrégation}
\usage{
controler_demandes(table_demandes)
}
\arguments{
\item{table_demandes}{data.frame avec colonnes :
\itemize{
\item id_demande : Identifiant unique
\item fonctions_agregations : Fonctions séparées par "-"
\item var_quanti : Variables séparées par "-"
\item unite : Unités séparées par "-"
\item var_croisement : Variables de croisement (optionnel)
\item var_croisement_relative : Variable pour calcul relatif (optionnel)
\item var_evolution : Variable d'évolution temporelle (optionnel)
\item type_output : "table" ou "graphique"
}}
}
\value{
data.frame des erreurs avec colonnes id_demande et raison
}
\description{
Vérifie la validité des demandes d'agrégation selon plusieurs critères :
\enumerate{
\item Cohérence des dimensions (fonctions/variables/unités)
\item Existence des fonctions et variables
\item Règles métier spécifiques
\item Cohérence des variables de croisement
}
}
\examples{
donnees <- creer_exemple_minimal()
erreurs_valides <- controler_demandes(donnees$table_demandes_valides)
erreurs_invalides <- controler_demandes(donnees$table_demandes_erreurs)

}
