% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_graphique_bar.R
\name{creer_graphique_bar}
\alias{creer_graphique_bar}
\title{Créer un graphique à barres avec ggplot2 et le sauvegarder automatiquement}
\usage{
creer_graphique_bar(
  data,
  var_x,
  var_y,
  var_fill,
  lab_x,
  lab_y,
  titre,
  labels_fill,
  soustitre = NULL,
  color_theme = c("#FF4858", "#1B7F79", "#00CCC0", "#72F2EB", "#747F7F"),
  param_position = "stack"
)
}
\arguments{
\item{data}{\code{data.frame} contenant lesgit push -u origin main données à représenter.}

\item{var_x}{Chaîne de caractères spécifiant le nom de la variable pour l'axe des abscisses dans \code{data}.}

\item{var_y}{Chaîne de caractères sgipécifiant le nom de la variable pour l'axe des ordonnées dans \code{data}.}

\item{var_fill}{Chaîne de caractères spécifiant le nom de la variable de remplissage (couleur) dans \code{data}.}

\item{lab_x}{Chaîne de caractères pour le label de l'axe des abscisses.}

\item{lab_y}{Chaîne de caractères pour le label de l'axe des ordonnées.}

\item{titre}{Chaîne de caractères pour le titre du graphique.}

\item{labels_fill}{Vecteur de chaînes de caractères pour les labels de la légende des couleurs.}

\item{soustitre}{(Optionnel) Chaîne de caractères pour le sous-titre du graphique. Par défaut : \code{NULL}.}

\item{color_theme}{Vecteur de codes couleurs hexadécimaux pour le graphique. Par défaut : palette de couleurs prédéfinie.}

\item{param_position}{Chaîne de caractères spécifiant la position des barres (\code{"stack"}, \code{"dodge"}, etc.). Par défaut : \code{"stack"}.}
}
\value{
Un objet \code{ggplot2} représentant le graphique à barres.
}
\description{
Cette fonction génère un graphique à barres à partir d'un jeu de données en utilisant ggplot2.
Le graphique est automatiquement sauvegardé dans un répertoire 'output' avec un nom de fichier
généré à partir du titre du graphique.
}
\details{
La fonction crée un graphique à barres en utilisant ggplot2 et le sauvegarde automatiquement dans un répertoire 'output'.
Le nom du fichier est généré à partir du titre du graphique en remplaçant les espaces par des underscores.

Les couleurs utilisées pour le remplissage sont déterminées par \code{color_theme}, et les labels de la légende sont spécifiés par \code{labels_fill}.
Si le répertoire 'output' n'existe pas, il sera créé automatiquement.
}
\examples{
\dontrun{
creer_graphique_bar(
  data = taux_global_long,
  var_x = "dep_EP",
  var_y = "percentage",
  var_fill = "region",
  lab_x = "Département",
  lab_y = "Pourcentage",
  titre = "Structure globale de détention",
  labels_fill = c("filiales 971", "filiales 972")
)
}

}
