% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_gerer_demandes.R
\name{ecrire_demande_sur_xls}
\alias{ecrire_demande_sur_xls}
\title{Écrire une table agrégée dans un fichier Excel}
\usage{
ecrire_demande_sur_xls(vecteur_demande, metadata = NULL)
}
\arguments{
\item{element_table_param}{Liste contenant la table et ses paramètres}
}
\value{
NULL (effet de bord : création fichier Excel)
}
\description{
Écrire une table agrégée dans un fichier Excel
}
