% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecrire_etude.R
\name{ecrire_etude}
\alias{ecrire_etude}
\title{Génère un rapport automatique à partir des métadonnées d'une étude}
\usage{
ecrire_etude(
  nom_etude,
  ollama = FALSE,
  model_name = "mistral-small",
  role = NULL
)
}
\arguments{
\item{nom_etude}{Nom de l'étude (chaîne de caractères)}

\item{ollama}{Utilisation d'Ollama pour la génération (logique, défaut FALSE)}

\item{model_name}{Nom du modèle Ollama (chaîne, défaut "mistral-small")}

\item{role}{Rôle utilisateur pour personnalisation (chaîne, optionnel)}
}
\value{
Génère un fichier Rmd et son rendu HTML dans le dossier output/
}
\description{
Génère un rapport automatique à partir des métadonnées d'une étude
}
\examples{
ecrire_etude("titanic", ollama = TRUE, role = "analyste")
}
