% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EP_FI_AG}
\alias{EP_FI_AG}
\title{Données de relations Entreprises-Filiales en Antilles-Guyane}
\format{
\subsection{EP_FI_AG}{

Un data frame avec 60 lignes et 13 colonnes:
\describe{
\item{annee}{Année d'observation (2022 ou 2023)}
\item{siren_EP}{Numéro SIREN de l'entreprise principale}
\item{denom_EP}{Dénomination de l'entreprise principale}
\item{dep_EP}{Département de l'entreprise principale (971, 972)}
\item{s4_EP}{Secteur d'activité de l'EP (Industrie, Services marchands)}
\item{redi_r310_EP}{Montant aléatoire pour l'EP (1000-5000)}
\item{siren_FI}{Identifiant de la filiale (FI1 à FI60)}
\item{denom_FI}{Dénomination de la filiale}
\item{dep_FI}{Département de la filiale (971-974)}
\item{s4_FI}{Secteur d'activité de la filiale}
\item{redi_r310_FI}{Montant aléatoire pour la filiale (500-2000)}
\item{poids}{Poids statistique (1)}
\item{dep_FI_relatif}{Position relative de la filiale (interieur, antilles-guyane, autres)}
}
}
}
\usage{
EP_FI_AG
}
\description{
Données de relations Entreprises-Filiales en Antilles-Guyane
}
\keyword{datasets}
