% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerer_demandes.R
\name{generer_figure_from_demande}
\alias{generer_figure_from_demande}
\title{Générer une figure à partir d'une table agrégée}
\usage{
generer_figure_from_demande(
  vecteur_demande,
  metadonnees_tables,
  ollama = FALSE,
  fonction_ask = NULL,
  metadonnees_etude
)
}
\arguments{
\item{vecteur_demande}{Vecteur nommé des paramètres de la demande}

\item{metadonnees_tables}{Liste des métadonnées des tables}

\item{ollama}{Booléen indiquant si utiliser Ollama}

\item{fonction_ask}{Fonction de requête optionnelle}

\item{metadonnees_etude}{Liste des métadonnées de l'étude}
}
\value{
Liste contenant la table agrégée, la réponse Ollama et les paramètres
}
\description{
Générer une figure à partir d'une table agrégée
}
