% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_simple_gerer_demandes.R
\name{generer_graphique_from_table}
\alias{generer_graphique_from_table}
\title{Générer un graphique en barres à partir d'une table agrégée}
\usage{
generer_graphique_from_table(element_table_param)
}
\arguments{
\item{element_table_param}{Liste contenant la table et ses paramètres}
}
\value{
Un objet ggplot2
}
\description{
Générer un graphique en barres à partir d'une table agrégée
}
