% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_agregations.R
\name{gerer_agregation_sum}
\alias{gerer_agregation_sum}
\title{Gérer l'agrégation avec la fonction sum}
\usage{
gerer_agregation_sum(
  table,
  var_croisement,
  var_croisement_relative = c(),
  var_quanti
)
}
\arguments{
\item{table}{Data.table contenant les données.}

\item{var_croisement}{Variables de croisement.}

\item{var_croisement_relative}{Variables de croisement relatives (optionnel).}

\item{var_quanti}{Variable quantitative à agréger.}
}
\value{
Un data.table avec les sous-totaux, totaux et parts.
}
\description{
Gérer l'agrégation avec la fonction sum
}
