% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_agregations.R
\name{get_rounding_function}
\alias{get_rounding_function}
\title{Obtenir une fonction d'arrondi selon l'unité spécifiée}
\usage{
get_rounding_function(unite)
}
\arguments{
\item{unite}{Character. L'unité d'arrondi souhaitée. Valeurs possibles :
"unite", "dixieme", "centieme", "millieme", "million"}
}
\value{
Une fonction qui prend un nombre et retourne sa valeur arrondie
selon l'unité spécifiée
}
\description{
Obtenir une fonction d'arrondi selon l'unité spécifiée
}
\examples{
f_arrondi <- get_rounding_function("dixieme")
f_arrondi(3.14159)  # Retourne 3.1

f_millions <- get_rounding_function("million")
f_millions(1234567)  # Retourne 1
}
