% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installer_package.R
\name{installer_package}
\alias{installer_package}
\title{Installer un package s'il n'est pas dejà installe}
\usage{
installer_package(pkg)
}
\arguments{
\item{pkg}{Un caractere unique indiquant le nom du package à installer.}
}
\value{
Aucun retour de valeur. Le package est installe si necessaire.
}
\description{
Cette fonction verifie si le package specifie est installe. Si ce n'est pas le cas, elle l'installe depuis CRAN.
}
\examples{
\dontrun{
installer_package("ggplot2")
}
}
