% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadonnees_etude}
\alias{metadonnees_etude}
\title{Métadonnées générales de l'étude EP-FI Antilles-Guyane}
\format{
Une liste imbriquée contenant:
\describe{
\item{nom}{Nom de l'étude}
\item{description}{Présentation générale}
\item{public_cible}{
\itemize{
\item type: Public visé
\item instructions_communication: Directives de communication
\item vocabulaire_adapte: Mapping des termes techniques
}
}
}
}
\usage{
metadonnees_etude
}
\description{
Métadonnées générales de l'étude EP-FI Antilles-Guyane
}
\examples{
metadonnees_etude$public_cible$vocabulaire_adapte$entreprise_profilée
}
\keyword{datasets}
