% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadonnees_role}
\alias{metadonnees_role}
\title{Métadonnées de rôle pour la communication}
\format{
Une liste avec 2 éléments principaux :
\describe{
\item{type}{Type de public cible}
\item{instructions_communication}{Liste contenant :
\itemize{
\item style : Style de communication à adopter
\item directives : Vecteur des règles de communication
\item vocabulaire_adapte : Liste des traductions de termes techniques
}
}
}
}
\usage{
metadonnees_role
}
\description{
Un jeu de données contenant les paramètres de communication adaptés
pour un public cible spécifique.
}
\keyword{datasets}
