% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadonnees_tables}
\alias{metadonnees_tables}
\title{Métadonnées des tables de l'étude EP-FI Antilles-Guyane}
\format{
Un data.frame contenant les descriptions des variables:
\describe{
\item{nom_etude}{Nom de l'étude}
\item{nom_table}{Nom de la table de données}
\item{nom_variable}{Identifiant technique de la variable}
\item{libelle}{Description complète de la variable}
\item{libelle_court}{Version abrégée du libellé}
\item{description}{Description détaillée de la variable}
}
}
\usage{
metadonnees_tables
}
\description{
Métadonnées des tables de l'étude EP-FI Antilles-Guyane
}
\keyword{datasets}
