% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_simple_gerer_demandes.R
\name{preparer_parametres_demande}
\alias{preparer_parametres_demande}
\title{Préparer les paramètres d'une demande de traitement}
\usage{
preparer_parametres_demande(vecteur_demande)
}
\arguments{
\item{vecteur_demande}{Un vecteur nommé contenant les paramètres de la demande}
}
\value{
Une liste structurée des paramètres traités
}
\description{
Préparer les paramètres d'une demande de traitement
}
