% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerer_demandes.R
\name{renvoyer_table_from_demande}
\alias{renvoyer_table_from_demande}
\title{Générer une table agrégée à partir d'une demande}
\usage{
renvoyer_table_from_demande(vecteur_demande)
}
\arguments{
\item{vecteur_demande}{Vecteur nommé des paramètres de la demande}

\item{metadata}{Liste des métadonnées pour la transformation des noms}
}
\value{
Liste contenant la table agrégée (table) et ses paramètres (params)
}
\description{
Générer une table agrégée à partir d'une demande
}
