% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_demandes_rmd}
\alias{table_demandes_rmd}
\title{Tables de demandes pour l'analyse des entreprises antillaises}
\format{
\subsection{table_demandes_rmd}{

Un data frame avec 20 lignes et 13 colonnes:
\describe{
\item{id_demande}{Identifiant unique de la demande}
\item{table}{Nom de la table source (EP_FI_AG)}
\item{condition}{Conditions de filtrage}
\item{var_croisement}{Variables de croisement pour l'analyse}
\item{var_croisement_relative}{Variable relative pour le croisement}
\item{var_evolution}{Variable d'évolution temporelle}
\item{var_quanti}{Variables quantitatives à analyser}
\item{unite}{Unité de mesure ('millier' ou 'unite')}
\item{fonctions_agregations}{Fonctions d'agrégation utilisées (sum, mean, max)}
\item{titre}{Titre de l'analyse}
\item{partie}{Section principale du rapport}
\item{sous_partie}{Sous-section du rapport}
\item{type_figure}{Type de visualisation ('kable' ou 'graphique')}
}
}
}
\source{
Données internes INSEE
}
\usage{
table_demandes_rmd
}
\description{
Tables de demandes pour l'analyse des entreprises antillaises
}
\keyword{datasets}
