% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerer_demandes.R
\name{transformer_noms_colonnes}
\alias{transformer_noms_colonnes}
\title{Transformer les noms de colonnes avec les métadonnées}
\usage{
transformer_noms_colonnes(noms_colonnes, nom_table, metadata)
}
\arguments{
\item{noms_colonnes}{Vecteur des noms de colonnes à transformer}

\item{nom_table}{Nom de la table dans les métadonnées}

\item{metadata}{Liste des métadonnées avec descriptions des variables}
}
\value{
Vecteur des noms de colonnes transformés
}
\description{
Transformer les noms de colonnes avec les métadonnées
}
