% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_creer_graphique_bar.R
\name{creer_graphique_bar}
\alias{creer_graphique_bar}
\title{Creer un graphique a barres avec ggplot2 et le sauvegarder automatiquement}
\usage{
creer_graphique_bar(
  data,
  var_x,
  var_y,
  var_fill,
  lab_x,
  lab_y,
  titre,
  labels_fill = NULL,
  soustitre = NULL,
  color_theme = c("#FF4858", "#1B7F79", "#00CCC0", "#72F2EB", "#747F7F"),
  param_position = "stack",
  save = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} contenant lesgit push -u origin main donnees a representer.}

\item{var_x}{Chaîne de caracteres specifiant le nom de la variable pour l'axe des abscisses dans \code{data}.}

\item{var_y}{Chaîne de caracteres sgipecifiant le nom de la variable pour l'axe des ordonnees dans \code{data}.}

\item{var_fill}{Chaîne de caracteres specifiant le nom de la variable de remplissage (couleur) dans \code{data}.}

\item{lab_x}{Chaîne de caracteres pour le label de l'axe des abscisses.}

\item{lab_y}{Chaîne de caracteres pour le label de l'axe des ordonnees.}

\item{titre}{Chaîne de caracteres pour le titre du graphique.}

\item{labels_fill}{Vecteur de chaînes de caracteres pour les labels de la legende des couleurs.}

\item{soustitre}{(Optionnel) Chaîne de caracteres pour le sous-titre du graphique. Par defaut : \code{NULL}.}

\item{color_theme}{Vecteur de codes couleurs hexadecimaux pour le graphique. Par defaut : palette de couleurs predefinie.}

\item{param_position}{Chaîne de caracteres specifiant la position des barres (\code{"stack"}, \code{"dodge"}, etc.). Par defaut : \code{"stack"}.}

\item{save}{booleen permettant de choisir si l'on veut sauvegarder ou non.}
}
\value{
Un objet \code{ggplot2} representant le graphique a barres.
}
\description{
Cette fonction genere un graphique a barres a partir d'un jeu de donnees en utilisant ggplot2.
Le graphique est automatiquement sauvegarde dans un repertoire 'output' avec un nom de fichier
genere a partir du titre du graphique.
}
\details{
La fonction cree un graphique a barres en utilisant ggplot2 et le sauvegarde automatiquement dans un repertoire 'output'.
Le nom du fichier est genere a partir du titre du graphique en remplaçant les espaces par des underscores.

Les couleurs utilisees pour le remplissage sont determinees par \code{color_theme}, et les labels de la legende sont specifies par \code{labels_fill}.
Si le repertoire 'output' n'existe pas, il sera cre automatiquement.
}
\examples{
\dontrun{
creer_graphique_bar(
  data = taux_global_long,
  var_x = "dep_EP",
  var_y = "percentage",
  var_fill = "region",
  lab_x = "Departement",
  lab_y = "Pourcentage",
  titre = "Structure globale de detention",
  labels_fill = c("filiales 971", "filiales 972")
)
}

}
