% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparer_prompt.R
\name{df_to_text}
\alias{df_to_text}
\title{Convert a data.frame to CSV text with line breaks}
\usage{
df_to_text(df)
}
\arguments{
\item{df}{A data.frame to convert to text}
}
\value{
A string containing the CSV representation of the data.frame
}
\description{
Convert a data.frame to CSV text with line breaks
}
\examples{
df <- data.frame(a = 1:3, b = letters[1:3])
df_to_text(df)
}
