% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_simple_ecrire_dans_xlsx.R
\name{ecrire_xls}
\alias{ecrire_xls}
\title{Écrire un Tableau avec Formatage et Titre dans un Fichier Excel}
\usage{
ecrire_xls(
  nom_fichier_xls,
  nom_onglet,
  table,
  titre,
  var_group_by,
  dir = "output/"
)
}
\arguments{
\item{nom_fichier_xls}{\code{character} : Le nom du fichier Excel à créer ou modifier (par exemple, "rapport.xlsx").}

\item{nom_onglet}{\code{character} : Le nom de l'onglet (feuille de calcul) où les données seront écrites.}

\item{table}{\code{data.frame} : Le tableau de données à écrire dans le fichier Excel.}

\item{titre}{\code{character} : Le titre à insérer au-dessus du tableau dans l'onglet Excel.}

\item{var_group_by}{\code{numeric} : Le nombre de premières colonnes à formater spécifiquement
(par exemple, pour des regroupements ou des croisements).}

\item{dir}{\code{character} : Le répertoire où le fichier Excel sera enregistré. Par défaut, "output/".}
}
\value{
\code{NULL}. La fonction écrit les données dans le fichier Excel spécifié.
}
\description{
Cette fonction écrit un tableau de données dans un fichier Excel en ajoutant un titre formaté
et en appliquant des styles de formatage au tableau. Elle gère également la création du
répertoire de sortie si nécessaire et permet d'ajouter des données à un onglet existant
en décalant les nouvelles données.
}
\details{
La fonction réalise les opérations suivantes :
\itemize{
\item Vérifie si le package \code{openxlsx} est installé et le charge.
\item Crée le répertoire de sortie s'il n'existe pas.
\item Charge un classeur Excel existant ou en crée un nouveau.
\item Détermine la position de départ pour écrire les nouvelles données en fonction
des données existantes dans l'onglet spécifié.
\item Ajoute ou sélectionne l'onglet spécifié.
\item Écrit le titre en gras et rouge au-dessus du tableau.
\item Appelle la fonction \code{\link{ecrire_tableau_formate}} pour écrire et formater le tableau.
\item Sauvegarde le classeur Excel.
}
}
\examples{
\dontrun{
library(openxlsx)

# Créer un exemple de tableau
donnees <- data.frame(
  Nom = c("Alice", "Bob", "Charlie"),
  Age = c(25, 30, 35),
  Score = c(85, 90, 95)
)

# Définir le titre et les variables de groupement
titre <- "Rapport des Scores"
var_group_by <- 2

# Utiliser la fonction pour écrire le tableau dans un fichier Excel
ecrire_xls(
  nom_fichier_xls = "rapport_scores.xlsx",
  nom_onglet = "Feuille1",
  table = donnees,
  titre = titre,
  var_group_by = var_group_by,
  dir = "output/"
)
}

}
