% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_prompt.R
\name{format_variables_metadata}
\alias{format_variables_metadata}
\title{Formater les métadonnées des variables}
\usage{
format_variables_metadata(metadonnees_tables, params)
}
\arguments{
\item{params}{Paramètres de configuration}

\item{metadata}{Liste des métadonnées des tables}
}
\value{
Chaîne formatée des descriptions des variables
}
\description{
Formater les métadonnées des variables
}
