% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_gerer_demandes.R
\name{generer_figure_from_demande}
\alias{generer_figure_from_demande}
\title{Générer un graphique en barres à partir d'une table agrégée}
\usage{
generer_figure_from_demande(
  vecteur_demande,
  metadonnees_tables,
  ollama = FALSE,
  fonction_ask = NULL,
  metadonnees_etude
)
}
\arguments{
\item{element_table_param}{Liste contenant la table et ses paramètres}
}
\value{
Un objet ggplot2
}
\description{
Générer un graphique en barres à partir d'une table agrégée
}
