% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerer_demandes.R
\name{gerer_une_demande}
\alias{gerer_une_demande}
\title{Gérer une demande d'agrégation}
\usage{
gerer_une_demande(numero_ligne, table_demandes)
}
\arguments{
\item{numero_ligne}{Un entier indiquant le numéro de la ligne dans \code{table_demandes} correspondant à la demande à traiter.}

\item{table_demandes}{Un \code{data.frame} contenant les demandes d'agrégation. Chaque ligne doit inclure les colonnes nécessaires telles que \code{table}, \code{var_croisement}, \code{var_croisement_relative}, \code{fonctions_agregations}, \code{condition}, et \code{var_quanti}.}
}
\value{
Un \code{data.frame} contenant le résultat de l'agrégation calculée.
}
\description{
Cette fonction traite une demande spécifique d'agrégation en récupérant les informations nécessaires depuis une table de demandes, puis en calculant l'agrégat en fonction des paramètres spécifiés.
}
\details{
La fonction effectue les étapes suivantes :
\enumerate{
\item Récupère les informations de la ligne spécifiée dans \code{table_demandes}.
\item Charge la table de données correspondante en utilisant \code{get()}.
\item Extrait les variables de croisement et les fonctions d'agrégation.
\item Évalue la condition fournie pour filtrer les données si nécessaire.
\item Calcule l'agrégat en appelant la fonction \code{calculer_agregat_sur_croisement()} avec les paramètres extraits.
}
}
\examples{
\dontrun{
# Exemple d'utilisation de gerer_une_demande
# Supposons que table_demandes soit un data.frame correctement formaté
resultat <- gerer_une_demande(numero_ligne = 1, table_demandes = table_demandes)
}

}
\seealso{
\code{\link{calculer_agregat_sur_croisement}}
}
