% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_prompt.R
\name{preparer_prompt}
\alias{preparer_prompt}
\title{Préparer un prompt pour l'analyse de données}
\usage{
preparer_prompt(table_agrege, metadonnees_tables, metadonnees_etude, params)
}
\arguments{
\item{metadonnees_tables}{Métadonnées des tables}

\item{metadonnees_etude}{Métadonnées de l'étude}

\item{table}{Table de données à analyser}
}
\value{
Chaîne de caractères contenant le prompt complet
}
\description{
Préparer un prompt pour l'analyse de données
}
