% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_simple_gerer_demandes.R
\name{renvoyer_table_from_demande}
\alias{renvoyer_table_from_demande}
\title{Générer une table agrégée à partir d'une demande}
\usage{
renvoyer_table_from_demande(vecteur_demande, metadata = NULL)
}
\arguments{
\item{vecteur_demande}{Un vecteur nommé contenant les paramètres de la demande}

\item{metadata}{Liste optionnelle des métadonnées pour la transformation des noms}
}
\value{
Une liste contenant la table agrégée et ses paramètres associés
}
\description{
Générer une table agrégée à partir d'une demande
}
