% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_simple_gerer_demandes.R
\name{transformer_noms_colonnes}
\alias{transformer_noms_colonnes}
\title{Transformer les noms de colonnes avec les métadonnées}
\usage{
transformer_noms_colonnes(noms_colonnes, metadata, nom_table)
}
\arguments{
\item{noms_colonnes}{Vecteur de noms de colonnes à transformer}

\item{metadata}{Liste des métadonnées contenant les descriptions des variables}

\item{nom_table}{Nom de la table dans les métadonnées}
}
\value{
Vecteur de noms de colonnes transformés
}
\description{
Transformer les noms de colonnes avec les métadonnées
}
