% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage_simple_gerer_demandes.R
\name{transformer_noms_colonnes}
\alias{transformer_noms_colonnes}
\title{Transformer les noms de colonnes avec les métadonnées}
\usage{
transformer_noms_colonnes(noms_colonnes, metadata, nom_table)
}
\arguments{
\item{noms_colonnes}{Un vecteur de noms de colonnes à transformer}

\item{metadata}{La liste des métadonnées contenant les descriptions des variables}

\item{nom_table}{Le nom de la table dans les métadonnées}
}
\value{
Un vecteur de noms de colonnes transformés
}
\description{
Transforme les noms de colonnes bruts en libellés plus compréhensibles en utilisant
les métadonnées et en gérant les suffixes spéciaux (_sum, _part, etc.).
}
